Fix contributed by 	@jobe451
Currently, the convert function passed on must return a Buffer in order to be effective. If it is not a buffer, the conversion is just not used. Please have a look at the screenshot below, you get to that finding when walking through the code.

Currently a user needs to pass-in a conversion function that returns buffer. However, for the user it is not obvious that a buffer-returning function is required. So I guess the better approach is to just also be able to cope with string-returning functions.